dird: move RunOnIncomingConnectInterval to separate files
- testability
- encapsulate functionality
